requirements.txt
fastapi
pydantic
firebase-admin
python-dotenv
elevenlabs
uvicorn
